Create PythonCode
n = number of bulbs in the problem
command will give the series of commands which describe the starting and ending bulbs in the range of bulbs. Ex: if n = 5 there are five bulbs. The command [1,4] will start on the first bulb and end on the fourth bulb.

the function toggles the bulbs on and off according to the command's range. [1,3] will turn [1] off, [2] on, [3] off, etc.

using the function we can find the end result of the bulbs after all commands have been executed.